facial recognition now only works if a picture is shown in webview and stops once a new picture gets loaded into it; a dialog asks the user whether the detection result was correct or not; TODO: save everything to a file

+++ app/src/main/java/com/example/ndk_opencv_androidstudio/face_detection/FaceDataRetriever.java
@@ -105,7 +105,7 @@ public class FaceDataRetriever extends Overlay.ViewTransformation {
         );
     }
 
-    public final float getSmilingProbability() {
+    public final double getSmilingProbability() {
         return smilingProbability;
     }
